Fix for cookie issues ,maybe, fix missing profile dir
On a fresh install, if the user doesn't configure any settings the
profile directory may not exist. This would cause a script error when
trying to save cookies. This adds a check.
In an attempt to avoid cookie issues, we now make a request to
'servlets/simpleconsole' before trying to resolve the url for a stream.