[new dep] Replace urllib2 with Requests 2.x
Requests is more concise and requires far less boilerplate.
urllib is still imported as it is used for internal param
passing. That will be removed eventually.

With this change, check_login logs in more than it needs to.
That will be fixed in a future patch that simplifies the
login/login_check process.