kill set_resolved_url(). Removed improper localization
Move logic/dialogs into appropriate area in skin code.

Removed localization for internal only keys. Makes it much easier to
debug/read.

Some of the "preferred version" code could be moved to a function, but I
want to wait on that until I have a cleaner idea of a new, restructured
way for game_common to efficiently return the information the skin
needs.