Rename week_enabled to "future" and invert meaning
week_enabled wasn't the best name, since it applied to games as well.
"future" is a bit more generic and accurate.

`ListItem.Property(future)` (when future isn't a valid property) returns false
anyway, so we don't need to set "future" to false when the game isn't in the
future (the most common scenario). This saves us a little bit of code too.