Move bitrate selection to skin; stream manifest returned to skin
This /should/ be split into multiple patches, but I didn't want to commit
pieces that resulted in a non-working state.

This /finally/ removes the need to import xbmcgui in game_common. It
also provides some significant cleanup. A big step forward for Issue

The real meat here is in two parts. 1) parse_request cleanup and 2)
bitrate selection

1) parse_request() no longer returns a url. Instead it returns the full
manifest (in dict form) which ends up at the skin layer. This could be
useful in the future, as duration info, codec info, etc is also passed
along, and might be helpful for added info in the skin or simply giving
the player helpful codec hints.

One bug-workaround was removed, handling for multiple httpservers.
GamePass looks to have cleaned up their returned XML a bit, so I want
to make sure it's needed before reintroducing this fix.

2) bitrate selection is a complicated thing. First of all, we have to
maintain a hardcoded list for the sake of the user's preference setting.
Secondly, for Live urls, we just guess and connect to a stream that we
anticipate (but are not certain) is there. However, with non-live games
and archived shows, we /do/ know each of the available bitrates for a
given stream. In those instances, we need to see if the user's preferred
bitrate is available, and if not, prompt the user with a list of bitrates
actually available in the manifest (rather than our hard coded internal list).

Localization was removed from ask_bitrate (the title remains localized). It
doesn't adapt well to the unknown nature of returned bitrates in the manifest.
The preference in addon settings /is/ still localized though. The only thing
of real value that wasn't ported is the resolution information (e.g. 1280x720).
This information is present in the manifest, so if we really care, we can pass
it to select_bitrate()/ask_bitrate() in the future.

Renamed functions:
  get_episode_url -> get_episode_manifest
  get_stream_url -> get_show_manifest

Removed functions:
  select_bitrate (in game_common, default.py's remains)

New functions:
  ask_bitrate (in default.py)