Updated/Replaced Dependency
The requests2 module is no longer offered by XBMC.

It used to be: requests == 1.1.0 and requests2 == 2.2.1
Now, they removed requests2 (which is a good idea), and requests == 2.3.0.

There is an "optional" tag for addon.xml requirements, but it doesn't have an
"either or" option. So... it seems to be best to just follow XBMC's play here.
Suddenly drop support for requests2 and make everyone install requests.

Close #121

---Alex