Bump version requirements (Helix+)
Bump xbmc.python to Helix, and xmltodict and requests to latest. The xmltodict
bump isn't strictly needed, but I figured I'd do it while bumping the others.
Requests bump ensures that no one is using the poorly renamed requests2 module,
which this commit also removes a check for.

Towards issue #149