Show all NFL Network Shows
Used #341 as a starting point.

The core problem is that Game Pass doesn't keep their seasons list for each show
up-to-date.

The "fix" is to loop over all shows, get all episodes for all shows, and use the
season info from each one.

This generates a lot of network traffic and slows down the initial init quite a
bit (varies from 1-15 seconds for me now). But, it's a fix.

However, not every episode has a season coded. Many contain "null". So, as a
fallback, the scheduled broadcast date is used. However, /that/ data apparently
is occasionally wrong... sigh (2018). That is a bug that needs to be fixed as
well. :-/

Lastly, since the season info isn't always available, we can always get the
season slug, so a hard-coded fallback is used.

If only we could submit bug reports to their API for returning faulty data...
:-/

Fixes #340
Close #341