Fix display of 'None' for focused games in game week view
The Gamepass API no longer provides the team city/state via the visitorCityState
and homeCityState elements. Both are set to 'null' -> the interface
displays None.

At least for past games and the current and upcoming game day we can retrieve
the state information from the video tags. Unfortunately, the API does not
provide this information for future games, so we default to the short display
version for those games.