Don't look up fixed bitrates anymore
It seems NFL Game Pass moved the app and the backend to adaptive streaming,
fixed bitrates are no longer supported. So Inputstream Adaptive is the
way to go forward and should be the default.

There's no need to look up URLs for fixed bitrates anymore, so the code now
returns the url to the best playlist.