Toggle display of current week's games
By default the add-on loads and displays the current week's games,
if the GamePass main selection is clicked. This could spoil someone,
if she/he would like to watch older games before jumping to the current
game week (e.g. during playoffs).

A new setting was added to disable the display of the current week's
games. By default, this option is enabled, so nothing changes to
the previous behavior.